requirements.txt
numpy
pandas
scikit-learn>=1.0
imbalanced-learn
xgboost
matplotlib
seaborn
scipy
pyyaml
joblib
shap
